feat: add `updateSubject` handler :construction:

+++ pages/api/subjects/[id].ts
@@ -4,11 +4,15 @@ import { connect } from "db";
 import { SubjectModel } from "db/models";
 import { formatApiError } from "utils/api";
 
-import type { DeleteSubjectData } from "types/api/subjects";
+import type {
+  DeleteSubjectData,
+  UpdateSubjectData,
+  UpdateSubjectRequestBody,
+} from "types/api/subjects";
 import type { ApiInternal, ApiInternalResponse } from "types/api";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-type Return = ApiInternalResponse<DeleteSubjectData>;
+type Return = ApiInternalResponse<DeleteSubjectData | UpdateSubjectData>;
 
 async function deleteSubject(_id: string) {
   await connect();
@@ -30,8 +34,32 @@ async function deleteSubject(_id: string) {
   return [result, statusCode] as const;
 }
 
+async function updateSubject(_id: string, data: UpdateSubjectRequestBody) {
+  await connect();
+  let [result, statusCode]: ApiInternal<UpdateSubjectData> = ["", 0];
+
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+  const res = await SubjectModel.findByIdAndUpdate(_id, data, {
+    fields: "_id",
+    runValidators: true,
+  }).lean();
+
+  [result, statusCode] = [
+    {
+      success: true,
+      data: {
+        success: !!res?._id,
+      },
+      message: ReasonPhrases.OK,
+    },
+    StatusCodes.OK,
+  ];
+
+  return [result, statusCode] as const;
+}
+
 export default async function handler(
-  { method = "", query }: NextApiRequest,
+  { body, method = "", query }: NextApiRequest,
   res: NextApiResponse<Return[0]>
 ) {
   const allow = ["DELETE"];
@@ -45,8 +73,14 @@ export default async function handler(
   ];
 
   try {
-    if (method === "DELETE" && typeof query.id === "string")
-      await deleteSubject(query.id);
+    if (typeof query.id === "string") {
+      if (method === "DELETE") await deleteSubject(query.id);
+      if (method === "PUT" && typeof body === "string")
+        await updateSubject(
+          query.id,
+          JSON.parse(body) as UpdateSubjectRequestBody
+        );
+    }
   } catch (error: any) {
     [result, statusCode] = [
       {
